+++ libs/frontend/features/src/lib/users/user-list/user-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnDestroy, OnInit } from '@angular/core';
 import { UserService } from '../user.service';
 import { IUser, IUserInfo } from '@avans-nx-expedition/shared/api';
 import { Subscription } from 'rxjs';
+import { AccountService } from '@avans-nx-expedition/frontend/account';
 
 @Component({
     selector: 'avans-nx-expedition-user-list',
@@ -11,8 +12,12 @@ import { Subscription } from 'rxjs';
 export class UserListComponent implements OnInit, OnDestroy {
     users: IUser[] | undefined = undefined;
     sub: Subscription = new Subscription();
+    loggedInUserId: string | null = '';
 
-    constructor(private userService: UserService) {}
+    constructor(
+        private userService: UserService,
+        private accountService: AccountService
+    ) {}
 
     ngOnInit(): void {
         console.log('UserListComponent.ngOnInit() aangeroepen');
@@ -27,23 +32,33 @@ export class UserListComponent implements OnInit, OnDestroy {
                 }
             )
         );
+
+        this.accountService.getLoggedInUserId().subscribe((userId) => {
+            this.loggedInUserId = userId;
+        });
     }
 
     deleteUser(userId: string): void {
+        if (this.loggedInUserId !== userId) {
+            console.error(`You cannot delete someone else's account.`);
+            return;
+        }
         console.log('Deleting user:', userId);
-        this.sub.add(
-            this.userService.deleteUser(userId).subscribe(
-                () => {
-                    console.log('User deleted successfully');
-                    this.users = this.users?.filter(
-                        (user) => user._id !== userId
-                    );
-                },
-                (error) => {
-                    console.error('Error deleting user:', error);
-                }
-            )
-        );
+        if (confirm('Are you sure you want to delete this activity?')) {
+            this.sub.add(
+                this.userService.deleteUser(userId).subscribe(
+                    () => {
+                        console.log('User deleted successfully');
+                        this.users = this.users?.filter(
+                            (user) => user._id !== userId
+                        );
+                    },
+                    (error) => {
+                        console.error('Error deleting user:', error);
+                    }
+                )
+            );
+        }
     }
 
     ngOnDestroy(): void {
